REFACTOR: Add processing and asset_requests modules
